Clarify what the proposed "5% noise" applies to
In the last Web-Advertising Business Group call, there was some confusion over what the noise applied to — the 3 bits of conversion metadata, or the reporting of the conversion event itself.  Added a bit of clarifying text, so that someone who just sees one of the two noise sections is directed to the other one as well.