Prevent registration of non-viewable impressions
Currently some ad tech players are sending impression events even if the ad has not been correctly seen. To address that, there a third party vendors relying on third party cookie to confirm that an impression has been viewed to do the attribution. It would be great to have this directly supported by the browser.