Allow attributionsrc requests to be made to untrustworthy origins
Without this, the handling of untrustworthy origins within a redirect
chain is inconsistent: The initial request is never made if the first
URL is untrustworthy, but intermediate untrustworthy origins are
silently ignored.

Note that we still require an origin to be trustworthy in order for its
registrations to be processed; this just affects whether any request is
made to untrustworthy ones.