Remove not- prefix from greased values and parameters
We retain the not- prefix to grease the dictionary keys, but use the
real keys to grease the dictionary values and parameters to ensure that
only the keys are interpreted by the recipient, not the values or
parameters. Removing the not- prefix in these contexts makes it less
likely for naive string operations to "work".