Remove JSON schemas
Unfortunately, JSON schema is not powerful enough to fully validate our
schemas with respect to the Full Flex proposal, and it is tedious to
maintain them alongside the more powerful custom JS-based validator.

The exact caveats for Full Flex are:

- Inability to ensure that trigger data values across trigger specs are
  disjoint
- Inability to ensure that summary buckets and report windows are
  increasing
- Inability to ensure that the first report window is greater than the
  start time

Additionally, we will likely want to retain the custom validator going
forward so that its output can be fed into the library that performs
privacy-related computations for Flex and shown in the web-based header
validator.