Prevent duplicate values in "obtain default effective windows" algorithm
By removing only elements that were greater than eventReportWindow, this
algorithm would have erroneously placed duplicate values in the
deadlines list if the eventReportWindow was exactly 2 days or 7 days. In
turn, this would have caused the returned windows list to contain 1
extra window.

This seems to have been an oversight in #742.