Make default event-level attributions per source a spec constant
Now that max_event_level_reports is configurable per source
registration, there seems to be little benefit for different
implementations to have different defaults here. Making this a constant
ensures interoperability between specs, simplifies logic for reporting
origins and implementers, and eliminates the only remaining parsing
behavior that the spec allows to vary in this way.