Require vendor-specific values for source/trigger validation
Proper validation relies more and more on the presence of all
vendor-specific values, so requiring this here simplifies the code,
especially for the forthcoming Full Flexible Event-Level changes.

If there are alternative vendor-specific values in the future, we will
expose them in the interactive header validator using a dropdown menu.