Perform event-report-window check after deduplication instead of before (#1092)
Event-level report deduplication is based only on whether a given
deduplication key has previously been seen for a given source. As such,
it makes more sense to perform deduplication as soon as both of these
values are known, instead of after the event-report-window check, which
is not special with respect to any of the existing checks that are
already performed after deduplication.

This change affects backwards compatibility in a minor way, in that the
"trigger-event-deduplicated" verbose debug report will now be sent
preferentially over the "trigger-event-report-window-not-started" and
"trigger-event-report-window-passed" verbose debug reports, but this
only affects triggers that are allowed to send verbose debug reports at
all and is only observable when both such failure conditions occur
together.

This changes makes the specification more readily compatible with the
Full Flexible Event Level proposal, as with that proposal, the
event-report-window check can only be performed after the trigger spec,
if any, has been found for the trigger's trigger data.