Reorganize report-delivery algorithms to prevent certain race conditions (#1105)
We remove reports from the cache prior to sending them to ensure that reports that are truly pending (and therefore subject to modification or deletion for various reasons) are not mistaken for ones that are in the process of being sent. This also allows us to remove the separate fields for original report time and report time, as while a report is in the report cache, its report time is never modified.