Restrict the number of report states per source (#1203)
Estimated channel capacity when the number of states is sufficiently high is effectively zero, and a randomised source response is effectively guaranteed. Additionally, a very high number of states coupled with the complexity of the flexible event feature can present challenges to performance. Altogether, user-agents may have cause to restrict the total number of report states per source.

Co-authored-by: Andrew Paseltiner <apaseltiner@google.com>
Co-authored-by: Nan Lin <80365263+linnan-github@users.noreply.github.com>