Remove redundant inclusion of pending source in scope records
The intent of this algorithm is that the pending source is never
rejected, but including it in the record list makes it seem like it can
be removed; that is only prevented by the "if selectedScopes does not
contain record[0]" check.

The pending source's scopes are already included in the selected scopes
set, so processing the pending source in the scopeRecords loop is
redundant.