Update AGGREGATION_SERVICE_TEE.md
Updating No Duplicate rule to say that the Shared Id can't be in more than one batch even though the reports can be shared.

With contribution Filtering, the filtering Id is included in the Shared ID generation. So, the reports can feature in different batches, but the Shared ID representing the disjoint data in them, can't be.